Add RichTextEditor bundle
Main bundle before: 421.5kB gzip
Main bundle after: 386.6kB gzip
New RichTextEditor bundle: 35.43kB gzip
